requirements.txt
crewai 
chainlit 
pdfminer.six
langchain
faiss-cpu
sentence-transformers
lightrag-hku 
ollama 
pdfplumber
textract



# ollama pull nomic-embed-text
# ollama pull mistral
